added project crud functions

+++ controllers/employee.js
@@ -1,5 +1,7 @@
 const { Employee } = require('../models');
-const { StatusCodes } = require('http-status-codes')
+const { NotFoundError } = require('../errors/not-found');
+const { BadRequestError } = require('../errors/bad-request');
+const { StatusCodes } = require('http-status-codes');
 
 const index = async (req, res) => {
     try {
@@ -15,7 +17,7 @@ const create = async (req, res) => {
         const emp = await Employee.create({ ...req.body });
         return res.status(StatusCodes.CREATED).json(emp);
     } catch (e) {
-        return res.json(e);
+        return res.status(StatusCodes.BAD_REQUEST).json({"message": e.message});
     }
 };
 
@@ -27,27 +29,29 @@ const read = async (req, res) => {
             }
         });
         if (!emp) {
-            return res.status(StatusCodes.NOT_FOUND).json({ "message": "employee not found" });
+            throw new NotFoundError("Employee Not Found");
         }
         res.status(StatusCodes.OK).json(emp);
     } catch (e) {
-        return res.json(e);
+        return res.status(StatusCodes.NOT_FOUND).json({"message": e.message});
     }
 };
 
 const update = async (req, res) => {
-    const id = req.params.id;
     try {
         const emp = await Employee.findOne({
             where:{
                 id: req.params.id
             }
         });
+        if(!emp){
+            throw new NotFoundError("Employee Not Found");
+        }
         await emp.set({...req.body});
         emp.save();
         return res.status(StatusCodes.OK).json(emp);
     } catch (e) {
-        return res.json(e);
+        return res.status(StatusCodes.BAD_REQUEST).json({"message" : e.message});
     }
 };
 
@@ -58,6 +62,9 @@ const destroy = async (req, res) => {
                 id: req.params.id,
             }
         });
+        if(!emp){
+            throw new NotFoundError("Employee not Found");
+        }
         await Employee.destroy({
             where: {
                 id: req.params.id,
@@ -65,11 +72,12 @@ const destroy = async (req, res) => {
         });
         return res.status(StatusCodes.OK).json(emp);
     } catch (e) {
-        return res.json(e);
+        return res.status(StatusCodes.NOT_FOUND).json({"message": e.message});
     }
 };
 
 const employeeController = {
+    index,
     create,
     read,
     update,
